Navigate to emotion details when pressing emotion name on intensity questions
The reasoning behind this is that if the user doesn't know what the
emotion is, then they need to learn. And right know the emotion
details is the best place to do that. It will contain a textual
description of the emotion in the future, and possibly some way to
relate the emotion to other emotions.